requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
cachetools==4.0.0         # via google-auth
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
google-api-core==1.16.0   # via google-cloud-core
google-auth==1.11.0       # via google-api-core
google-cloud-bigquery==1.23.1
google-cloud-core==1.2.0  # via google-cloud-bigquery
google-resumable-media==0.5.0  # via google-cloud-bigquery
googleapis-common-protos==1.51.0  # via google-api-core
idna==2.8                 # via requests
protobuf==3.11.2          # via google-api-core, google-cloud-bigquery, googleapis-common-protos
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pytz==2019.3              # via google-api-core
requests==2.22.0          # via google-api-core
rsa==4.0                  # via google-auth
six==1.14.0               # via google-api-core, google-auth, google-cloud-bigquery, google-resumable-media, protobuf
urllib3==1.25.8           # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
